backend/requirements.txt
Flask==2.2.3
flask-cors==3.0.10
python-dotenv==1.0.0
PyJWT==2.6.0
Flask-SQLAlchemy==3.0.3
sqlalchemy==2.0.7
mysql-connector-python==8.0.32
pymysql==1.0.2
werkzeug==2.2.3

# S3互換API用
boto3==1.26.84
botocore>=1.29.84,<1.30.0
